strcmp.c: add qsort benchmark
This benchmark generates an array of random strings and then
sorts it lexicographically using strcmp() and qsort().  This
should be a bit more realistic of a benchmark then just
comparing strings with themselves.

The alphabet size is reduced to 16 for more realistic conditions,
ensuring that at least some strings share prefixes.

Sponsored by:	The FreeBSD Foundation